# Optimize metadata reading for extension methods