--- a/src/Compiler/AbstractIL/ilread.fs
+++ b/src/Compiler/AbstractIL/ilread.fs
@@ -2144,6 +2144,11 @@ and typeDefReader ctxtH : ILTypeDefStored =
             else
                 let mutable attrIdx = attrsStartIdx
 
+                let looksLikeSystemAssembly =
+                    ctxt.fileName.EndsWith("System.Runtime.dll")
+                    || ctxt.fileName.EndsWith("mscorlib.dll")
+                    || ctxt.fileName.EndsWith("netstandard.dll")
+
                 while attrIdx <= attrsEndIdx && not containsExtensionMethods do
                     let mutable addr = ctxt.rowAddr TableNames.CustomAttribute attrIdx
                     seekReadHasCustomAttributeIdx ctxt mdv &addr |> ignore
@@ -2152,8 +2157,11 @@ and typeDefReader ctxtH : ILTypeDefStored =
 
                     let name =
                         if attrTypeIndex.tag = cat_MethodDef then
-                            let _, (_, nameIdx, namespaceIdx, _, _, _) = seekMethodDefParent ctxt attrCtorIdx
-                            readBlobHeapAsTypeName ctxt (nameIdx, namespaceIdx)
+                            if not looksLikeSystemAssembly then
+                                ""
+                            else
+                                let _, (_, nameIdx, namespaceIdx, _, _, _) = seekMethodDefParent ctxt attrCtorIdx
+                                readBlobHeapAsTypeName ctxt (nameIdx, namespaceIdx)
                         else
                             let mutable addr = ctxt.rowAddr TableNames.MemberRef attrCtorIdx
                             let mrpTag = seekReadMemberRefParentIdx ctxt mdv &addr