… docs/release-notes/.FSharp.Compiler.Service/8.0.400.md
… src/Compiler/CodeGen/IlxGen.fs
… tests/FSharp.Compiler.Service.Tests/FSharp.Compiler.Service.SurfaceArea.netstandard20.debug.bsl
… tests/FSharp.Compiler.Service.Tests/FSharp.Compiler.Service.SurfaceArea.netstandard20.release.bsl