Currently, finding out whether a type in an IL assembly has `Extension` attribute requires reading all attributes from the metadata and constructing a representation of `ILAttribute` for each one. This process is time-consuming and involves unnecessary work. It would be beneficial to have a more efficient method for determining if a type can contain extension methods.

Furthermore, `ILTypeDef` constructor is a public API but lacks the ability to pass lazy computation of custom attributes. In contrast, the internal constructor in FCS supports this feature.

This PR proposes:
- The ability to pass lazy computation of custom attributes to `ILTypeDef` public constructor
- The ability to specify the presence of `Extension` attribute in `ILTypeDef` (and some refactoring around `isKnownToBeAttribute` flag)
- Default implementation for lightweight detection of `Extension` attribute presence

This solution will allow for more efficient filtering of `ILTypeDef` that definitely does not contain extension methods, both for users of the public API and within the FCS implementation.