@DedSec256 2024-06-19
@T-Gro, @psfinaki, I think it's ready for review. 

Lazy evaluation of attributes (our own implementation inside ReSharper.FSharp) and the presence of extension methods flag allowed us to eliminate unnecessary attribute calculations (from 30 seconds to 1 second), reduce the `ILTypeDef` creation time by 30 seconds, and the overall analysis time of the F# file inside the Rider monorepository from 70 seconds to 40 seconds.